Initial Commit
This commit is the initial commit of deciders-rs.

Completely implements pretty much the entirety of the example code from
Jérémie Chassaing's Aggregate Composition talk.

I'm not completely sold on using generics for everything like I
currently have but it does pretty much force the user to not store any
state (unless they decide to use global statics...) which is nice. The
syntax can be a bit heavy though.

License is TODO.